requirements.txt
pandas>=1.5.0
numpy>=1.23.0
matplotlib>=3.6.0
scipy>=1.9.0
pyyaml>=6.0
pytest>=7.2.0
pytest-cov>=4.0.0